refactor(inference-optimizer): move model-config helpers to shared leaf util
Extract _load_model_config_dict / _config_architectures / _model_is_gemma2
into inference_optimizer/model_config_utils.py (stdlib-only leaf module).
cli re-exports them (names unchanged, all internal call sites untouched) and
_workload_envs now imports _model_is_gemma2 from the util instead of cli,
removing the executor->cli back-import / circular-import risk. No behavior
change; verified re-export identity and real-model detection under
/wekafs/models.
